tg_chanel/bot_naeb/requirements.txt
aiogram==3.0.1
python-dotenv